Rename to bamboo-rs-*
