+++ bamboo-rs-cli/.gitignore
