+++ bamboo-rs-cli/test_script/lipmaa.sh
