+++ bamboo-rs-cli/test_script/test.sh
