+++ bamboo-rs-core-test/Cargo.toml
@@ -1,15 +1,15 @@
 [package]
-name = "bamboo-core-test"
+name = "bamboo-rs-core-test"
 version = "1.0.0"
 authors = ["Piet Geursen <pietgeursen@gmail.com>"]
 edition = "2018"
 
 [dependencies]
-bamboo-core = {path = "../bamboo-core"}
+bamboo-rs-core = {path = "../bamboo-rs-core"}
 
 [dev-dependencies]
-bamboo-core = {path = "../bamboo-core"}
-bamboo-log = {path = "../bamboo-log"}
+bamboo-rs-core = {path = "../bamboo-rs-core"}
+bamboo-rs-log = {path = "../bamboo-rs-log"}
 arrayvec = { version = "0.5.1", default-features = false}
 serde = { version = "1.0.94", default-features = false , features = ["derive"] }
 serde_derive = { version = "1.0.94", default-features = false }
