+++ bamboo-rs-core-test/fuzz/fuzz_targets/fuzz_entry_add.rs
