+++ bamboo-rs-core/libbamboo.h
