+++ bamboo-rs-core/src/entry/decode.rs
