+++ bamboo-rs-core/src/entry/encode.rs
