+++ bamboo-rs-core/src/entry/mod.rs
