+++ bamboo-rs-core/src/entry/publish.rs
