+++ bamboo-rs-core/src/entry/verify.rs
