+++ bamboo-rs-core/src/entry/verify_batch.rs
