+++ bamboo-rs-core/src/error/mod.rs
