+++ bamboo-rs-core/src/util/hex_serde.rs
