+++ bamboo-rs-core/src/util/mod.rs
