+++ bamboo-rs-core/src/yamf_hash.rs
