+++ bamboo-rs-log/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "bamboo-log"
+name = "bamboo-rs-log"
 version = "1.0.0"
 authors = ["Piet Geursen <pietgeursen@gmail.com>"]
 edition = "2018"
@@ -7,7 +7,7 @@ edition = "2018"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-bamboo-core = {path = "../bamboo-core"}
+bamboo-rs-core = {path = "../bamboo-rs-core"}
 lipmaa-link = "0.1.1"
 
 [dev-dependencies]
