+++ bamboo-rs-wasm/.appveyor.yml
