+++ bamboo-rs-wasm/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "bamboo-wasm"
+name = "bamboo-rs-wasm"
 version = "0.4.0"
 authors = ["Piet Geursen <pietgeursen@gmail.com>"]
 edition = "2018"
@@ -13,7 +13,7 @@ default = ["wee_alloc", "console_error_panic_hook"]
 [dependencies]
 arrayvec = { version = "0.5", default-features = false}
 wasm-bindgen = { version = "0.2.68", features = ["serde-serialize"] }
-bamboo-core = {path = "../bamboo-core"}
+bamboo-rs-core = {path = "../bamboo-rs-core"}
 ed25519-dalek = { version = "1.0.1", default-features = false, features = ["u64_backend", "alloc", "serde"] }
 rand = { version = "0.7.3", features = ["wasm-bindgen"] }
 serde = { version = "1.0.90", features = ["derive"] }
