+++ bamboo-rs-wasm/src/utils.rs
